fix: enabled the fixture for clean-up before/after tests also to integration tests
